+++ client/src/App.tsx
@@ -4,8 +4,6 @@ import {
   BrowserRouter,
   Route,
   Routes,
-  useLocation,
-  useNavigate,
   useNavigation,
   useParams,
   useResolvedPath,
@@ -23,10 +21,9 @@ import { useAuthStore } from "./stores/authStore";
 import { supabaseClient } from "./utils/supabase-auth";
 import { setAuthCookie } from "./api/api";
 import SniffersPage from "./pages/sniffers/SniffersPage";
-import { a11yDark, CopyBlock } from "react-code-blocks";
-import { TfiSettings } from "react-icons/tfi";
-import { AiOutlinePlayCircle } from "react-icons/ai";
 import { LandingPage } from "./LandingPage";
+import { SharkioDocsGettingStartedPage } from "./pages/docs/SharkioDocsGettingStartedPage";
+import { SharkioDocsSetupPage } from "./pages/docs/SharkioDocsSetupPage";
 
 function App(): React.JSX.Element {
   const { mode } = useThemeStore();
@@ -144,233 +141,4 @@ function App(): React.JSX.Element {
   );
 }
 
-const SharkioDocsSetupPage = () => {
-  return (
-    <SharkioDocsPageTemplate
-      title={"Setup Guide"}
-      subTitle={
-        "Learn how to set up Sharkio and start using it to develop your APIs."
-      }
-    >
-      <SharkioDocsSection title="Sign Up" sectionNumber={1}>
-        Head to the{" "}
-        <a href="https://sharkio.dev" className="hover:text-blue-200">
-          Sharkio website.
-        </a>
-        <img
-          src="/login.png"
-          alt=""
-          className="w-full rounded-lg h-96 border-2 border-border-color mt-4"
-        />
-      </SharkioDocsSection>
-
-      <SharkioDocsSection title="Generate an API Token" sectionNumber={2}>
-        <p>
-          Access the{" "}
-          <a
-            href="https://sharkio.dev/api-keys"
-            className="hover:text-blue-200"
-          >
-            API Keys page
-          </a>{" "}
-          by clicking on your profile picture and generate a token.
-        </p>
-        <img
-          src="/apiKeys.png"
-          alt=""
-          className="w-full rounded-lg h-96 border-2 border-border-color mt-4"
-        />
-      </SharkioDocsSection>
-
-      <SharkioDocsSection title="Install Sharkio-cli" sectionNumber={3}>
-        Install the Sharkio-cli.
-        <div className="flex flex-col w-full mt-2">
-          <CopyBlock
-            language="bash"
-            text={`npm i -g sharkio-cli`}
-            showLineNumbers={false}
-            theme={a11yDark}
-            wrapLines={true}
-            codeBlock
-          />
-        </div>
-      </SharkioDocsSection>
-
-      <SharkioDocsSection title="Sharkio Login" sectionNumber={4}>
-        Login to Sharkio-cli, the session will be saved in your home folder.
-        <div className="flex flex-col w-full my-2">
-          <CopyBlock
-            language="bash"
-            text={`sharkio login`}
-            showLineNumbers={false}
-            theme={a11yDark}
-            wrapLines={true}
-            codeBlock
-          />
-        </div>
-        {"You'll"} be prompted to enter your email and the precious token{" "}
-        {"you've "}secured.
-      </SharkioDocsSection>
-    </SharkioDocsPageTemplate>
-  );
-};
-
-const SharkioDocsGettingStartedPage = () => {
-  return (
-    <SharkioDocsPageTemplate
-      title={"Getting Started"}
-      subTitle={"Get started with Sharkio in 4 easy steps."}
-    >
-      <SharkioDocsSection title={"Start Your API"} sectionNumber={1}>
-        Start your API service on your local machine.
-      </SharkioDocsSection>
-
-      <SharkioDocsSection title="Create a Sniffer" sectionNumber={2}>
-        Create a sniffer (=Proxy) to monitor your API traffic by running the
-        following command:
-        <div className="flex flex-col w-full my-2">
-          <CopyBlock
-            language="bash"
-            text={`sharkio create sniffer -p <port> -n <name>`}
-            showLineNumbers={false}
-            theme={a11yDark}
-            wrapLines={true}
-            codeBlock
-          />
-        </div>
-        Port: The port your API is running on.{"\n"}
-        Name: The name of your sniffer.
-        <div className="text-lg font-serif mt-2">
-          {"You'll"} be given a unique subdomain for your sniffer. This is where
-          you will redirect your traffic to.
-        </div>
-      </SharkioDocsSection>
-
-      <SharkioDocsSection title="Start Your Sniffer" sectionNumber={3}>
-        Start your sniffer by running the following command:
-        <div className="flex flex-col w-full my-2">
-          <CopyBlock
-            language="bash"
-            text={`sharkio start sniffer`}
-            showLineNumbers={false}
-            theme={a11yDark}
-            wrapLines={true}
-            codeBlock
-          />
-        </div>
-      </SharkioDocsSection>
-
-      <SharkioDocsSection
-        title="Redirect Other Services to Subdomain"
-        sectionNumber={4}
-      >
-        Redirect your all the services that communicate with your API to the new
-        subdomain.{"\n"}
-        Now you can open the Sharkio dashboard and start building faster.
-      </SharkioDocsSection>
-    </SharkioDocsPageTemplate>
-  );
-};
-
-type SharkioDocsSectionProps = {
-  title: string;
-  children: React.ReactNode;
-  sectionNumber?: number;
-};
-
-const SharkioDocsSection = ({
-  title,
-  children,
-  sectionNumber,
-}: SharkioDocsSectionProps) => {
-  return (
-    <section className="flex flex-col mb-8">
-      <div className="flex w-full flex-row mb-2">
-        <div className="flex text font-serif bg-blue-200 px-2 rounded-lg border-blue-200 border-2 bg-opacity-40 mr-4">
-          {sectionNumber ? `${sectionNumber}` : 1}
-        </div>
-        <div className="text-xl font-bold whitespace-pre-line">{title}</div>
-      </div>
-      <div className="text-lg font-serif mb-4 text-[#717171] whitespace-pre-line">
-        {children}
-      </div>
-    </section>
-  );
-};
-
-type SharkioDocsNavigationProps = {
-  navigationItems: { title: string; path: string; icon: React.ReactNode }[];
-};
-
-const SharkioDocsNavigation = ({
-  navigationItems,
-}: SharkioDocsNavigationProps) => {
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  const onClick = (path: string) => {
-    navigate(path);
-  };
-
-  return (
-    <div className="flex-col w-1/5 border-r bg-secondary border-border-color hidden sm:flex sm:w-1/4 lg:w-1/5">
-      {navigationItems.map(({ title, path, icon }) => (
-        <div
-          key={title}
-          className={`flex text-lg font-serif text-center items-center border-b border-border-color hover:cursor-pointer hover:bg-tertiary 
-          
-            `}
-          onClick={() => onClick(path)}
-        >
-          <div className="flex flex-1 active:scale-95 px-4 py-2 items-center">
-            <div
-              className={`${location.pathname === path ? "text-blue-300" : ""}`}
-            >
-              {icon}
-            </div>
-            {title}
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-};
-
-type SharkioDocsPageTemplateProps = {
-  title: string;
-  subTitle: string;
-  children: React.ReactNode;
-};
-
-const SharkioDocsPageTemplate = ({
-  children,
-  title,
-  subTitle,
-}: SharkioDocsPageTemplateProps) => {
-  const navigationItems = [
-    {
-      title: "Getting Started",
-      path: routes.DOCS_GETTINGS_STARTED,
-      icon: <TfiSettings className="w-6 h-6 mr-2" />,
-    },
-    {
-      title: "Setup",
-      path: routes.DOCS_SETUP,
-      icon: <AiOutlinePlayCircle className="w-6 h-6 mr-2" />,
-    },
-  ];
-
-  return (
-    <div className="flex flex-1 flex-row">
-      <SharkioDocsNavigation navigationItems={navigationItems} />
-      <div className="p-4 flex flex-col flex-1 ">
-        <div className="flex text-4xl font-serif ">{title}</div>
-        <div className="flex w-full border-b border-border-color my-6"></div>
-        <div className="text-lg font-serif mb-4 text-[#717171]">{subTitle}</div>
-
-        <div className="flex flex-col flex-1 p-4">{children}</div>
-      </div>
-    </div>
-  );
-};
 export default App;
