+++ client/src/LandingPage.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { useNavigate } from "react-router-dom";
 import { routes } from "./constants/routes";
