--- client/src/components/config-card/EditProxyButton.tsx
@@ -1,41 +1,0 @@
-import { Edit, Save } from "@mui/icons-material";
-import { ConfigButton } from "./ConfigButton";
-
-export const EditProxyButton = ({
-  onEdit,
-  onSave,
-  disabled,
-  isEditing,
-  isLoading,
-}: {
-  onEdit: () => void;
-  onSave: () => void;
-  disabled?: boolean;
-  isEditing: boolean;
-  isLoading?: boolean;
-}) => {
-  return (
-    <>
-      {isEditing && (
-        <ConfigButton
-          tooltip={"Save Proxy"}
-          onClick={onSave}
-          disabled={disabled}
-          isLoading={isLoading}
-        >
-          <Save color={disabled ? "disabled" : "info"} />
-        </ConfigButton>
-      )}
-      {!isEditing && (
-        <ConfigButton
-          tooltip={"Edit Proxy"}
-          onClick={onEdit}
-          disabled={disabled}
-          isLoading={isLoading}
-        >
-          <Edit color={disabled ? "disabled" : "info"} />
-        </ConfigButton>
-      )}
-    </>
-  );
-};
