--- client/src/components/config-card/config-card.module.scss
@@ -1,42 +1,0 @@
-.container {
-  padding: 20px;
-}
-.sniffersContainer {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  column-gap: 10px;
-  row-gap: 10px;
-}
-
-.snifferTitle {
-  margin-bottom: 15px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-}
-.inputs {
-  .snifferTitle {
-    align-self: center;
-    text-transform: capitalize;
-    cursor: pointer;
-  }
-}
-
-.addSnifferBtn {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 10px 0px;
-}
-
-.addSnifferBtn > * {
-  cursor: pointer;
-  text-transform: none;
-}
-
-.titleBar {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 30px;
-}
