--- client/src/pages/config/config.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import { ConfigCard } from "../../components/config-card/config-card";
-
-export const Config: React.FC = () => {
-  return <ConfigCard />;
-};
