--- client/src/pages/home/home.module.scss
@@ -1,9 +1,0 @@
-.container > *:first-child {
-  height: 300px;
-  margin-bottom: 30px;
-  overflow: auto;
-}
-
-.container > *:last-child {
-  height: 400px;
-}
