+++ client/src/pages/sniffers/SniffersSideBar.tsx
@@ -47,7 +47,7 @@ export const SniffersSideBar = ({
   };
 
   return (
-    <div className="relative min-w-[240px] border-r-[0.1px] border-border-color">
+    <div className="relative min-w-[240px] border-r-[0.1px] border-border-color bg-secondary">
       <div className="flex h-16 items-center p-2 mb-2 justify-between">
         <div className="text-white text-xl font-bold">Sniffers</div>
         <AiOutlinePlus
